moved sample to test
