+++ disparity_view/__init__.py
@@ -1,4 +1,5 @@
 from .view import view_npy
+from .view import depth_overlay
 from .depth_to_normal import DepthToNormalMap
 from .zed_info import (
     get_width_height_fx_fy_cx_cy,
